admin/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-mysql==4.15.0
djangorestframework==3.15.2
mysqlclient==2.2.6
pika==1.3.2
sqlparse==0.5.2
tzdata==2024.2
psycopg2-binary
